refactor(android): split large files to remove TooManyFunctions suppresses
Extract DayCell + DayCellDots → DayCell.kt (CalendarRoute.kt: 13→9 fns)
Extract backup logic → CalendarBackupHandler (CalendarViewModel: 12→9 fns)
Extract chore dialogs → ChoreTemplateDialogs.kt (TemplatesRoute: 13→9 fns)

Also fixes DayCellDots parameter type: List<Int> → List<String>.
