+++ android/app/src/main/java/com/daynest/android/feature/calendar/CalendarBackupHandler.kt
@@ -1,0 +1,80 @@
+package com.daynest.android.feature.calendar
+
+import com.daynest.android.data.today.PlannedItemCreateDto
+import com.daynest.android.data.today.PlannedTodayItemDto
+import com.daynest.android.data.today.TodayRepository
+import java.time.Instant
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.async
+import kotlinx.coroutines.awaitAll
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.update
+import kotlinx.coroutines.launch
+
+internal class CalendarBackupHandler(
+    private val scope: CoroutineScope,
+    private val todayRepository: TodayRepository,
+    private val uiState: MutableStateFlow<CalendarUiState>,
+    private val onRefresh: () -> Unit,
+) {
+    fun exportMonthBackup(onReady: (PlannedItemBackupDto) -> Unit) {
+        val current = uiState.value as? CalendarUiState.Content ?: return
+        val startDate = current.displayMonth.withDayOfMonth(1).toString()
+        val endDate = current.displayMonth.withDayOfMonth(current.displayMonth.lengthOfMonth()).toString()
+        scope.launch {
+            val result = todayRepository.listPlannedItems(startDate, endDate)
+            result.onSuccess { items ->
+                onReady(
+                    PlannedItemBackupDto(
+                        source = "daynest",
+                        schemaVersion = 1,
+                        exportedAt = Instant.now().toString(),
+                        items = items.map { it.toBackupItem() },
+                    ),
+                )
+            }
+        }
+    }
+
+    fun importBackup(items: List<PlannedItemCreateDto>) {
+        scope.launch {
+            val results =
+                items.chunked(5).flatMap { batch ->
+                    batch
+                        .map { item -> async { todayRepository.createPlannedItem(item) } }
+                        .awaitAll()
+                }
+            val imported = results.count { it.isSuccess }
+            val failed = results.count { it.isFailure }
+            uiState.update { current ->
+                if (current is CalendarUiState.Content) {
+                    current.copy(backupMessage = CalendarBackupMessage.ImportComplete(imported, failed))
+                } else {
+                    current
+                }
+            }
+            onRefresh()
+        }
+    }
+
+    fun updateBackupMessage(message: CalendarBackupMessage) {
+        uiState.update { current ->
+            if (current is CalendarUiState.Content) {
+                current.copy(backupMessage = message)
+            } else {
+                current
+            }
+        }
+    }
+}
+
+private fun PlannedTodayItemDto.toBackupItem() =
+    PlannedItemBackupItemDto(
+        title = title,
+        plannedFor = plannedFor,
+        notes = notes,
+        moduleKey = moduleKey,
+        recurrenceHint = recurrenceHint,
+        linkedSource = linkedSource,
+        linkedRef = linkedRef,
+    )
