+++ android/app/src/main/java/com/daynest/android/feature/calendar/CalendarRoute.kt
@@ -1,9 +1,7 @@
-@file:Suppress("ktlint:standard:function-naming", "FunctionNaming", "TooManyFunctions")
+@file:Suppress("ktlint:standard:function-naming", "FunctionNaming")
 
 package com.daynest.android.feature.calendar
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -20,7 +18,6 @@ import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.Button
 import androidx.compose.material3.Card
 import androidx.compose.material3.CircularProgressIndicator
@@ -34,7 +31,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.pluralStringResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.platform.LocalContext
@@ -60,10 +56,6 @@ import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 
 private const val DAYS_IN_WEEK = 7
-private const val ITEM_TYPE_ROUTINE = "routine"
-private const val ITEM_TYPE_CHORE = "chore"
-private const val ITEM_TYPE_MEDICATION = "medication"
-private const val ITEM_TYPE_PLANNED = "planned"
 
 private val plannedItemBackupJson =
     Json {
@@ -434,90 +426,6 @@ private fun MonthGrid(
                     }
                 }
             }
-        }
-    }
-}
-
-@Composable
-private fun DayCell(
-    dayNum: Int,
-    routines: Int,
-    chores: Int,
-    medications: Int,
-    planned: Int,
-    isSelected: Boolean,
-    isToday: Boolean,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val bgColor =
-        when {
-            isSelected -> MaterialTheme.colorScheme.primary
-            isToday -> MaterialTheme.colorScheme.primaryContainer
-            else -> androidx.compose.ui.graphics.Color.Transparent
-        }
-    val textColor =
-        when {
-            isSelected -> MaterialTheme.colorScheme.onPrimary
-            isToday -> MaterialTheme.colorScheme.onPrimaryContainer
-            else -> MaterialTheme.colorScheme.onSurface
-        }
-    val dotTypes =
-        remember(routines, chores, medications, planned) {
-            buildList {
-                if (routines > 0) add(ITEM_TYPE_ROUTINE)
-                if (chores > 0) add(ITEM_TYPE_CHORE)
-                if (medications > 0) add(ITEM_TYPE_MEDICATION)
-                if (planned > 0) add(ITEM_TYPE_PLANNED)
-            }
-        }
-
-    Box(
-        modifier =
-            modifier
-                .aspectRatio(1f)
-                .padding(2.dp)
-                .clip(CircleShape)
-                .background(bgColor)
-                .clickable(onClick = onClick),
-        contentAlignment = Alignment.Center,
-    ) {
-        Column(horizontalAlignment = Alignment.CenterHorizontally) {
-            Text(
-                text = dayNum.toString(),
-                style = MaterialTheme.typography.bodySmall,
-                color = textColor,
-            )
-            if (dotTypes.isNotEmpty()) {
-                DayCellDots(dotTypes = dotTypes, isSelected = isSelected)
-            }
-        }
-    }
-}
-
-@Composable
-private fun DayCellDots(dotTypes: List<Int>, isSelected: Boolean) {
-    val colorScheme = MaterialTheme.colorScheme
-    Row(horizontalArrangement = Arrangement.spacedBy(2.dp)) {
-        dotTypes.forEach { typeIndex ->
-            val dotColor =
-                if (isSelected) {
-                    colorScheme.onPrimary
-                } else {
-                    when (typeIndex) {
-                        ITEM_TYPE_ROUTINE -> colorScheme.primary
-                        ITEM_TYPE_CHORE -> colorScheme.secondary
-                        ITEM_TYPE_MEDICATION -> colorScheme.tertiary
-                        else -> colorScheme.outline
-                    }
-                }
-            Box(
-                modifier =
-                    Modifier
-                        .size(3.dp)
-                        .clip(CircleShape)
-                        .background(dotColor),
-            )
         }
     }
 }
